requirements.txt
numpy
Pillow
streamlit
torch==1.10.1+cpu
torchvision==0.11.2+cpu
